failing tests are fixed
SVN-Revision: 31156

+++ contrib/brl/bbas/bocl/tests/test_kernel.cl
@@ -1,9 +1,9 @@
 __kernel 
 void
-test_multiply( __global float* input, 
-               __global float* output)
+test_multiply( __global int* input, 
+               __global int* output)
 {
   int gid = get_global_id(0)*get_global_size(0) + get_global_id(1);
-  float in    = input[gid]; 
+  int in    = input[gid]; 
   output[gid] = in * in; 
 }
